updating

+++ sdks/java/io/hcatalog/src/test/java/org/apache/beam/sdk/io/hcatalog/HCatalogIOTest.java
@@ -17,15 +17,15 @@
  */
 package org.apache.beam.sdk.io.hcatalog;
 
-import static org.apache.beam.sdk.io.hcatalog.HCatalogIOTestUtils.TEST_DATABASE;
-import static org.apache.beam.sdk.io.hcatalog.HCatalogIOTestUtils.TEST_FILTER;
-import static org.apache.beam.sdk.io.hcatalog.HCatalogIOTestUtils.TEST_RECORDS_COUNT;
-import static org.apache.beam.sdk.io.hcatalog.HCatalogIOTestUtils.TEST_TABLE;
-import static org.apache.beam.sdk.io.hcatalog.HCatalogIOTestUtils.buildHCatRecords;
-import static org.apache.beam.sdk.io.hcatalog.HCatalogIOTestUtils.getConfigPropertiesAsMap;
-import static org.apache.beam.sdk.io.hcatalog.HCatalogIOTestUtils.getExpectedRecords;
-import static org.apache.beam.sdk.io.hcatalog.HCatalogIOTestUtils.getReaderContext;
-import static org.apache.beam.sdk.io.hcatalog.HCatalogIOTestUtils.insertTestData;
+import static org.apache.beam.sdk.io.hcatalog.test.HCatalogIOTestUtils.TEST_DATABASE;
+import static org.apache.beam.sdk.io.hcatalog.test.HCatalogIOTestUtils.TEST_FILTER;
+import static org.apache.beam.sdk.io.hcatalog.test.HCatalogIOTestUtils.TEST_RECORDS_COUNT;
+import static org.apache.beam.sdk.io.hcatalog.test.HCatalogIOTestUtils.TEST_TABLE;
+import static org.apache.beam.sdk.io.hcatalog.test.HCatalogIOTestUtils.buildHCatRecords;
+import static org.apache.beam.sdk.io.hcatalog.test.HCatalogIOTestUtils.getConfigPropertiesAsMap;
+import static org.apache.beam.sdk.io.hcatalog.test.HCatalogIOTestUtils.getExpectedRecords;
+import static org.apache.beam.sdk.io.hcatalog.test.HCatalogIOTestUtils.getReaderContext;
+import static org.apache.beam.sdk.io.hcatalog.test.HCatalogIOTestUtils.insertTestData;
 import static org.hamcrest.Matchers.containsInAnyOrder;
 import static org.hamcrest.Matchers.containsString;
 import static org.hamcrest.Matchers.isA;
@@ -40,9 +40,14 @@ import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
+import org.apache.beam.sdk.coders.Coder;
 import org.apache.beam.sdk.io.BoundedSource;
+import org.apache.beam.sdk.io.hadoop.WritableCoder;
 import org.apache.beam.sdk.io.hcatalog.HCatalogIO.BoundedHCatalogSource;
+import org.apache.beam.sdk.io.hcatalog.test.EmbeddedMetastoreService;
 import org.apache.beam.sdk.options.PipelineOptions;
 import org.apache.beam.sdk.options.PipelineOptionsFactory;
 import org.apache.beam.sdk.testing.PAssert;
@@ -51,12 +56,16 @@ import org.apache.beam.sdk.testing.TestPipeline;
 import org.apache.beam.sdk.transforms.Create;
 import org.apache.beam.sdk.transforms.DoFn;
 import org.apache.beam.sdk.transforms.ParDo;
+import org.apache.beam.sdk.transforms.Watch;
 import org.apache.beam.sdk.util.UserCodeException;
 import org.apache.beam.sdk.values.PCollection;
+import org.apache.beam.vendor.guava.v20_0.com.google.common.collect.ImmutableList;
 import org.apache.hadoop.hive.metastore.api.NoSuchObjectException;
 import org.apache.hadoop.hive.ql.CommandNeedRetryException;
+import org.apache.hive.hcatalog.data.DefaultHCatRecord;
 import org.apache.hive.hcatalog.data.HCatRecord;
 import org.apache.hive.hcatalog.data.transfer.ReaderContext;
+import org.joda.time.Duration;
 import org.junit.AfterClass;
 import org.junit.BeforeClass;
 import org.junit.ClassRule;
@@ -67,9 +76,12 @@ import org.junit.rules.TemporaryFolder;
 import org.junit.rules.TestRule;
 import org.junit.rules.TestWatcher;
 import org.junit.runner.Description;
+import org.junit.runner.RunWith;
+import org.junit.runners.JUnit4;
 import org.junit.runners.model.Statement;
 
 /** Test for HCatalogIO. */
+@RunWith(JUnit4.class)
 public class HCatalogIOTest implements Serializable {
   private static final PipelineOptions OPTIONS = PipelineOptionsFactory.create();
 
@@ -93,6 +105,9 @@ public class HCatalogIOTest implements Serializable {
                 prepareTestData();
               } else if (description.getAnnotation(NeedsEmptyTestTables.class) != null) {
                 reCreateTestTable();
+              } else if (description.getAnnotation(NeedsEmptyTestTablesForUnboundedReads.class)
+                  != null) {
+                reCreateTestTableForUnboundedReads();
               }
               base.evaluate();
             }
@@ -107,6 +122,11 @@ public class HCatalogIOTest implements Serializable {
   @Target({ElementType.METHOD})
   private @interface NeedsTestData {}
 
+  /** Use this annotation to setup complete test data(table populated with unbounded records). */
+  @Retention(RetentionPolicy.RUNTIME)
+  @Target({ElementType.METHOD})
+  private @interface NeedsEmptyTestTablesForUnboundedReads {}
+
   /** Use this annotation to setup test tables alone(empty tables, no records are populated). */
   @Retention(RetentionPolicy.RUNTIME)
   @Target({ElementType.METHOD})
@@ -119,14 +139,16 @@ public class HCatalogIOTest implements Serializable {
 
   @AfterClass
   public static void shutdownEmbeddedMetastoreService() throws Exception {
-    service.executeQuery("drop table " + TEST_TABLE);
-    service.close();
+    if (service != null) {
+      service.executeQuery("drop table " + TEST_TABLE);
+      service.close();
+    }
   }
 
   /** Perform end-to-end test of Write-then-Read operation. */
   @Test
   @NeedsEmptyTestTables
-  public void testWriteThenReadSuccess() throws Exception {
+  public void testWriteThenReadSuccess() {
     defaultPipeline
         .apply(Create.of(buildHCatRecords(TEST_RECORDS_COUNT)))
         .apply(
@@ -158,9 +180,59 @@ public class HCatalogIOTest implements Serializable {
     readAfterWritePipeline.run();
   }
 
+  private Map<String, String> getPartitions() {
+    Map<String, String> partitions = new HashMap<>();
+    partitions.put("load_date", "2019-05-14T23:28:04.425Z");
+    partitions.put("product_type", "1");
+    return partitions;
+  }
+
+  /** Perform end-to-end test of Write-then-Read operation. */
+  @Test
+  @NeedsEmptyTestTablesForUnboundedReads
+  public void testWriteThenUnboundedReadSuccess() throws Exception {
+
+    defaultPipeline
+        .apply(Create.of(buildHCatRecords(TEST_RECORDS_COUNT)))
+        .apply(
+            HCatalogIO.write()
+                .withConfigProperties(getConfigPropertiesAsMap(service.getHiveConf()))
+                .withDatabase(TEST_DATABASE)
+                .withTable(TEST_TABLE)
+                .withPartition(getPartitions())
+                .withBatchSize(512L));
+    defaultPipeline.run();
+    final ImmutableList<String> partitions = ImmutableList.of("load_date", "product_type");
+    final PCollection<HCatRecord> data =
+        readAfterWritePipeline
+            .apply(
+                "ReadData",
+                HCatalogIO.read()
+                    .withConfigProperties(getConfigPropertiesAsMap(service.getHiveConf()))
+                    .withDatabase(TEST_DATABASE)
+                    .withPartitionCols(partitions)
+                    .withTable(TEST_TABLE)
+                    .withPollingInterval(Duration.millis(15000))
+                    .withTerminationCondition(Watch.Growth.afterTotalOf(Duration.millis(60000))))
+            .setCoder((Coder) WritableCoder.of(DefaultHCatRecord.class));
+
+    final PCollection<String> output =
+        data.apply(
+            ParDo.of(
+                new DoFn<HCatRecord, String>() {
+                  @ProcessElement
+                  public void processElement(ProcessContext c) {
+                    c.output(c.element().get(0).toString());
+                  }
+                }));
+
+    PAssert.that(output).containsInAnyOrder(getExpectedRecords(TEST_RECORDS_COUNT));
+    readAfterWritePipeline.run();
+  }
+
   /** Test of Write to a non-existent table. */
   @Test
-  public void testWriteFailureTableDoesNotExist() throws Exception {
+  public void testWriteFailureTableDoesNotExist() {
     thrown.expectCause(isA(UserCodeException.class));
     thrown.expectMessage(containsString("org.apache.hive.hcatalog.common.HCatException"));
     thrown.expectMessage(containsString("NoSuchObjectException"));
@@ -175,7 +247,7 @@ public class HCatalogIOTest implements Serializable {
 
   /** Test of Write without specifying a table. */
   @Test
-  public void testWriteFailureValidationTable() throws Exception {
+  public void testWriteFailureValidationTable() {
     thrown.expect(IllegalArgumentException.class);
     thrown.expectMessage("withTable() is required");
     HCatalogIO.write()
@@ -185,7 +257,7 @@ public class HCatalogIOTest implements Serializable {
 
   /** Test of Write without specifying configuration properties. */
   @Test
-  public void testWriteFailureValidationConfigProp() throws Exception {
+  public void testWriteFailureValidationConfigProp() {
     thrown.expect(IllegalArgumentException.class);
     thrown.expectMessage("withConfigProperties() is required");
     HCatalogIO.write().withTable("myowntable").expand(null);
@@ -193,7 +265,7 @@ public class HCatalogIOTest implements Serializable {
 
   /** Test of Read from a non-existent table. */
   @Test
-  public void testReadFailureTableDoesNotExist() throws Exception {
+  public void testReadFailureTableDoesNotExist() {
     defaultPipeline.apply(
         HCatalogIO.read()
             .withConfigProperties(getConfigPropertiesAsMap(service.getHiveConf()))
@@ -204,7 +276,7 @@ public class HCatalogIOTest implements Serializable {
 
   /** Test of Read without specifying configuration properties. */
   @Test
-  public void testReadFailureValidationConfig() throws Exception {
+  public void testReadFailureValidationConfig() {
     thrown.expect(IllegalArgumentException.class);
     thrown.expectMessage("withConfigProperties() is required");
     HCatalogIO.read().withTable("myowntable").expand(null);
@@ -212,7 +284,7 @@ public class HCatalogIOTest implements Serializable {
 
   /** Test of Read without specifying a table. */
   @Test
-  public void testReadFailureValidationTable() throws Exception {
+  public void testReadFailureValidationTable() {
     thrown.expect(IllegalArgumentException.class);
     thrown.expectMessage("withTable() is required");
     HCatalogIO.read()
@@ -266,9 +338,22 @@ public class HCatalogIOTest implements Serializable {
     SourceTestUtils.assertSourcesEqualReferenceSource(unSplitSource.get(0), splits, OPTIONS);
   }
 
-  private void reCreateTestTable() throws CommandNeedRetryException {
+  private void reCreateTestTable() {
     service.executeQuery("drop table " + TEST_TABLE);
     service.executeQuery("create table " + TEST_TABLE + "(mycol1 string, mycol2 int)");
+  }
+
+  private void reCreateTestTableForUnboundedReads() throws CommandNeedRetryException {
+    service.executeQuery("drop table " + TEST_TABLE);
+    service.executeQuery(
+        "create table "
+            + TEST_TABLE
+            + "(mycol1 string, mycol2 int)  "
+            + "partitioned by (load_date string, product_type string)");
+    service.executeQuery(
+        "ALTER TABLE "
+            + TEST_TABLE
+            + " ADD PARTITION (load_date='2019-05-14T23:28:04.425Z', product_type='1')");
   }
 
   private void prepareTestData() throws Exception {
